Refactored the remaining classes into sub-modules
