+++ src/directions/index.ts
@@ -1,0 +1,7 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export * from "./defines";
+export * from "./directions_renderer";
+export * from "./directions_service";
+export * from "./distance_matrix_service";
