+++ src/places/index.ts
@@ -1,4 +1,11 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export * from "./placeTypes";
+export * from "./autocomplete_service";
+export * from "./autocomplete";
+export * from "./opening_hours";
+export * from "./place";
+export * from "./place_conversion";
+export * from "./place_types";
+export * from "./places_service";
+export * from "./searchbox";
